[Feat] : filter로 수정
