+++ src/main/java/com/mogak/spring/config/SecurityConfig.java
@@ -1,7 +1,8 @@
 package com.mogak.spring.config;
 
-import com.mogak.spring.jwt.JwtInterceptor;
+//import com.mogak.spring.jwt.JwtInterceptor;
 //import com.mogak.spring.login.JwtTokenFilter;
+import com.mogak.spring.jwt.JwtTokenFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -18,7 +19,7 @@ import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 @RequiredArgsConstructor
 public class SecurityConfig {
 
-//    private final JwtTokenFilter jwtTokenFilter;
+    private final JwtTokenFilter jwtTokenFilter;
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception {
@@ -27,22 +28,19 @@ public class SecurityConfig {
                 .csrf().disable()
                 .formLogin().disable()
                 .cors().disable()
-                .build();
-                /*
-                .authorizeRequests()
-                .antMatchers("/h2-console/*").permitAll()
-                .antMatchers("**").permitAll() // 우선 모든 권한 허용
+//                .authorizeRequests()
+//                .antMatchers("/h2-console/*").permitAll()
+//                .antMatchers("**").permitAll() // 우선 모든 권한 허용
 //                .antMatchers(HttpMethod.POST,"/api/v1/**").authenticated()
                 // 모든 post 요청을 인증된 사용자인지 순서 중요. authenticated 🡪 인증된 사용자인지 확인
                 // .antMatchers("/api/**").authenticated() // 다른 api는 인증 필요
-                .and()
+                //.and()
                 .sessionManagement()
                 .sessionCreationPolicy(SessionCreationPolicy.STATELESS) // jwt 사용하는 경우 사용
                 .and()
                 .addFilterBefore(jwtTokenFilter, UsernamePasswordAuthenticationFilter.class)
                 //UserNamePasswordAuthenticationFilter 적용하기 전에 JWTTokenFilter를 적용 하라는 뜻.
                 .build();
-                 */
     }
 
 }
