+++ src/main/java/com/mogak/spring/config/WebConfig.java
@@ -1,7 +1,7 @@
 package com.mogak.spring.config;
 
 import com.amazonaws.HttpMethod;
-import com.mogak.spring.jwt.JwtInterceptor;
+//import com.mogak.spring.jwt.JwtInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -15,7 +15,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 public class WebConfig implements WebMvcConfigurer {
 
-    private final JwtInterceptor jwtInterceptor;
+//    private final JwtInterceptor jwtInterceptor;
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
@@ -29,18 +29,18 @@ public class WebConfig implements WebMvcConfigurer {
                 );
     }
 
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(jwtInterceptor)
-                .addPathPatterns("/api/**")
-                .excludePathPatterns(
-                        "/api/apple/login",
-                        "/api/apple/refresh",
-                        "/api/apple/logout",
-                        "/api/users/join",
-                        "/api/users/{nickname}/verify",
-                        "/swagger-ui/index.html"
-
-                );
-    }
+//    @Override
+//    public void addInterceptors(InterceptorRegistry registry) {
+//        registry.addInterceptor(jwtInterceptor)
+//                .addPathPatterns("/api/**")
+//                .excludePathPatterns(
+//                        "/api/apple/login",
+//                        "/api/apple/refresh",
+//                        "/api/apple/logout",
+//                        "/api/users/join",
+//                        "/api/users/{nickname}/verify",
+//                        "/swagger-ui/index.html"
+//
+//                );
+//    }
 }
